Don't wait poll interval to perform the first scrape
When the exporter first starts up it currently waits for one poll interval before fetching the first set of data. This means that during a restart of the exporter in production it may be possible to lose one poll interval of data. The initial delay also makes local testing slower than necessary. 

Also makes a couple of minor refactors in preparation for upcoming config changes:
* Move config type to a separate file
* Add wrapper type for interacting with cloudwatch labels