Fix race condition when fetching resources (#4)
Currently when fetching resources an empty array for the current list of resources is initialized and is saved in place of the existing list. If this list is accessed by one of the metrics go routines before it has been fully populated then that metric will not be populated for the missing resources.

This diff adds locking and changes the new array to only be saved once fully populated to prevent such data races from occurring.